Added foundation for Connect4

+++ index.js
@@ -17,10 +17,18 @@ const client = new Discord.Client()
 const authKey = require ("./auth.json")
 const packFile = require ("./package.json")
 const config = require ("./config.json")
+const commands = require ("./commands/commands.json")
+
+// Import the command router
+//const route = require ("./commands/router.js")
 
 // Variables
 let cmdPrefix = config.prefix
-var dateNow = Date(Date.now()).toString()
+var adminRole = "571780045924728833"
+
+function dateNow() {
+	return Date(Date.now())
+}
 
 // Rich Presence config thingy
 client.on('ready', () => {
@@ -44,6 +52,12 @@ client.on('message', (receivedMessage) => {
 	}
 })
 
+// Shut the bot down
+function botExit() {
+	console.warn("This isnt graceful but fuck it right?")
+	process.exit(0)
+}
+
 // Function handler
 function processCommand(receivedMessage) {
 	let fullCommand = receivedMessage.content.substr(2) // Remove prefix
@@ -55,9 +69,13 @@ function processCommand(receivedMessage) {
 	console.log("Command received: " + primaryCommand)
 	console.log("Arguments: " + arguments)
 
+	primaryCommand = primaryCommand.toLowerCase();
+
 	////////
 	// MOVE ALL TO SEPERATE FILES
 	////////
+
+	//route.routeCmd()
 
 	// Help command
 	if (primaryCommand == "help") {
@@ -74,6 +92,10 @@ function processCommand(receivedMessage) {
 	// Info
 	else if (primaryCommand == "info") {
 		infoCmd(receivedMessage)
+	}
+	// launch game manager
+	else if (primaryCommand == "play") {
+		playCmd(arguments, receivedMessage)
 	}
 	// Unknown command was entered
 	else {
@@ -96,8 +118,8 @@ function pingCmd(arguments, receivedMessage) {
 		.setColor(0xEF59D1)
 		.setTitle("Sapphire Bot Service Status")
 		.addField("__**Current Status**__", "Online")
-		.setThumbnail("https://i.imgur.com/u4DGhJP.png")
-		.setFooter(dateNow);
+		.setThumbnail("https://i.imgur.com/XHG5FTQ.png")
+		.setFooter(dateNow());
 	
 	// Now display the thingy
 	receivedMessage.channel.send(pingRE)
@@ -105,32 +127,69 @@ function pingCmd(arguments, receivedMessage) {
 	console.log("DEBUG: pingCMD has completed")
 }
 
+// Information dialoge
 function infoCmd(receivedMessage) {
 	receivedMessage.channel.send("**Sapphire Bot Info** \nCreated by: *Arek 'AGDeveloper' Kwapis* \nBot Version: *" + packFile.version + "*")
+}
+
+// Game Manager
+function playCmd(arguments, receivedMessage) {
+	if (arguments.length == 0) {
+		receivedMessage.channel.send("ADD PANEL HERE!")
+	}
+	else if (arguments[0] == "c4" || arguments[0] == "connect4" || arguments[0] == "connectfour") {
+		client.destroy() // Close the client
+		const connectFour = require("./commands/games/c4.js")
+		connectFour.c4Cmd()
+	}
 }
 
 // Debug
 function debugCmd(arguments, receivedMessage) {
 	if (arguments.length == 0) {
-		receivedMessage.channel.send("Debug Usage Feature WIP!")
+		receivedMessage.channel.send("No args passed to `debug`")
 	}
 	else {
+		// Restart
 		if (arguments == "restart") {
-			receivedMessage.channel.send("Restarting client, please wait...")
-			client.destroy() // Close client
-			console.log("DEBUG: Destroyed client")
-			client.login(authKey.token)
-			console.log("DEBUG: Logged into Discord on ", dateNow)
-			// Send message on REAL ready
-			client.on('ready', () => {
-				receivedMessage.channel.send("Restart has completed sucesfully!")
-				console.log("DEBUG: App is ready // called from the debug restart command")
-			})
-			console.log("DEBUG: debugCMD with Restart arg has completed")
+			let n = receivedMessage.member._roles.includes(adminRole)
+
+			if (n == 1) {
+				receivedMessage.channel.send("Restarting client, please wait...")
+				client.destroy() // Close client
+				console.log("DEBUG: Destroyed client")
+				client.login(authKey.token)
+				console.log("DEBUG: Logged into Discord on ", dateNow())
+				// Send message on REAL ready
+				client.on('ready', () => {
+					receivedMessage.channel.send("Restart has completed sucesfully!")
+					console.log("DEBUG: App is ready // called from the debug restart command")
+				})
+				console.log("DEBUG: debugCMD with Restart arg has completed")
+			}
+			else {
+				receivedMessage.channel.send("You do not have permission to use this command!")
+			}
+		}
+		// Shut down process
+		else if (arguments == "shdown") {
+			let n = receivedMessage.member._roles.includes(adminRole)
+
+			if (n == 1) {
+				receivedMessage.channel.send("Shutting down, please wait...")
+				botExit()
+			}
+			else {
+				receivedMessage.channel.send("You do not have permission to use this command!")
+			}
+		}
+		// Invalid command
+		else {
+			receivedMessage.channel.send("Invalid argument passed!")
 		}
 	}
 }
 
 // Login to client
 client.login(authKey.token)
-console.log("Logged into discord on ", dateNow)
+console.log("Logged into discord on", dateNow())
